+++ lib/vagrant/action/general/validate.rb
@@ -6,12 +6,11 @@ module Vagrant
       class Validate
         def initialize(app, env)
           @app = app
-          @env = env
         end
 
         def call(env)
-          @env[:vm].config.validate!(@env[:vm].env) if !@env.has_key?("validate") || @env["validate"]
-          @app.call(@env)
+          env[:vm].config.validate!(env[:vm].env) if !env.has_key?("validate") || env["validate"]
+          @app.call(env)
         end
       end
     end